refactor: modularize quality monitoring
