+++ src/services/quality/__init__.py
@@ -29,11 +29,9 @@ from .assurance_engine import (
 )
 
 # Quality Monitoring
-from .quality_monitor import (
-    QualityMonitor,
-    QualityAlertManager,
-    QualityTrendAnalyzer
-)
+from .data_gathering import QualityMonitor
+from .analysis import QualityAnalyzer
+from .alerting import QualityAlertManager
 
 # Quality Validation
 from .quality_validator import (
@@ -63,8 +61,8 @@ __all__ = [
     
     # Quality Monitoring
     "QualityMonitor",
+    "QualityAnalyzer",
     "QualityAlertManager",
-    "QualityTrendAnalyzer",
     
     # Quality Validation
     "QualityValidator",
